backend/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
networkx==3.2.1
matplotlib==3.8.2
anthropic==0.8.1
GitPython==3.1.41
PyGithub==2.1.1